Change price data type from integer to float

+++ db/migrate/20190520155621_add_price_to_stations.rb
@@ -1,5 +1,5 @@
 class AddPriceToStations < ActiveRecord::Migration[5.2]
   def change
-    add_column :stations, :price, :integer
+    add_column :stations, :price, :float
   end
 end
